requirements.txt
colorama==0.4.6
imbalanced-learn==0.13.0
imblearn==0.0
joblib==1.4.2
numpy==2.2.0
nvidia-nccl-cu12==2.23.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
sklearn-compat==0.1.3
threadpoolctl==3.5.0
tzdata==2024.2
xgboost==2.1.3
